move render queue funcs into own file
