+++ gridsome/lib/app/__tests__/pages.spec.js
@@ -1,6 +1,5 @@
 const path = require('path')
 const App = require('../App')
-const { createRenderQueue } = require('../pages')
 const { BOOTSTRAP_PAGES } = require('../../utils/constants')
 
 test('create page', async () => {
@@ -103,112 +102,6 @@ test('allways include a /404 page', async () => {
   const notFound = app.pages.findPage({ path: '/404' })
 
   expect(notFound.path).toEqual('/404')
-})
-
-test('generate render queue', async () => {
-  const app = await createApp(function plugin (api) {
-    api.loadSource(async store => {
-      const posts = api.store.addContentType({ typeName: 'Post', route: '/post/:id' })
-      const movies = api.store.addContentType({ typeName: 'Movie' })
-
-      for (let i = 1; i <= 3; i++) {
-        posts.addNode({ id: String(i), fields: { author: '2' }})
-      }
-
-      movies.addNode({
-        id: '1',
-        path: '/movie/one',
-        fields: {
-          posts: [
-            store.createReference('Post', '1')
-          ]
-        }
-      })
-
-      movies.addNode({
-        id: '2',
-        path: '/movie/two',
-        fields: {
-          posts: [
-            store.createReference('Post', '1'),
-            store.createReference('Post', '2')
-          ]
-        }
-      })
-
-      movies.addNode({
-        id: '3',
-        path: '/movie/three',
-        fields: {
-          posts: [
-            store.createReference('Post', '1')
-          ]
-        }
-      })
-    })
-
-    api.createPages(async ({ store, createPage, graphql }) => {
-      const posts = store.getContentType('Post')
-
-      createPage({
-        path: '/about',
-        component: './__fixtures__/DefaultPage.vue'
-      })
-
-      createPage({
-        path: '/blog',
-        component: './__fixtures__/PagedPage.vue',
-        context: { perPage: 2 }
-      })
-
-      for (let i = 1; i <= 3; i++) {
-        const node = posts.getNode(String(i))
-
-        createPage({
-          route: '/post/:id',
-          path: node.path,
-          component: './__fixtures__/DefaultTemplate.vue',
-          context: node
-        })
-      }
-
-      const movies = await graphql(`query {
-        allMovie {
-          edges {
-            node {
-              id
-              path
-            }
-          }
-        }
-      }`)
-
-      for (const edge of movies.data.allMovie.edges) {
-        createPage({
-          path: edge.node.path,
-          component: './__fixtures__/MovieTemplate.vue',
-          context: edge.node
-        })
-      }
-    })
-  })
-
-  const queue = createRenderQueue([], app)
-  const paths = queue.map(entry => entry.path)
-
-  expect(paths).toEqual(expect.arrayContaining([
-    '/about',
-    '/movie/three',
-    '/movie/two',
-    '/movie/one',
-    '/404',
-    '/blog',
-    '/blog/2',
-    '/post/1',
-    '/post/1/2',
-    '/post/2',
-    '/post/3'
-  ]))
 })
 
 async function createApp (plugin) {
